fix/add a 'component' verification condition for CSS styles of a Modal.js
