+++ src/components/BurgerIngredientsCard/BurgerIngredientsCard.js
@@ -2,14 +2,10 @@ import { CurrencyIcon, Counter } from '@ya.praktikum/react-developer-burger-ui-c
 import styles from './burgerIngredientsCard.module.css';
 import { burgerIngredientType, countType } from '../../utils/prop-types';
 import { useDrag } from 'react-dnd';
-import { useDispatch } from 'react-redux';
-import { SHOW_ITEM_DETAILS } from '../../services/actions/modalDetails';
 import { Link, useLocation } from 'react-router-dom';
 import { INGREDIENTS } from '../../constants/constants';
 
 export default function BurgerIngredientsCard({ ingridient, count }) {
-
-  const dispatch = useDispatch();
   const location = useLocation();
 
   const [{isDrag}, dragRef] = useDrag({
@@ -22,19 +18,11 @@ export default function BurgerIngredientsCard({ ingridient, count }) {
 
   const hover = isDrag ? styles.onHover : '';
 
-  const openModal = () => {
-    dispatch({
-    type: SHOW_ITEM_DETAILS,
-    item: ingridient
-    })
-  }
-
   return(
     <Link to={`${INGREDIENTS}/${ingridient._id}`} state={{ background: location }}
       ref={dragRef} className={ `${ styles.element } ${hover}` }>
       {count && <Counter count={count} size="default" extraClass="m-1" />}
       <img className={ styles.pic }
-        onClick={openModal}
         src={ingridient.image}
         alt="Картинка. Вид ингридиента">
       </img>
