--- src/components/Ingredient/ingredient.module.css
@@ -1,77 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  align-content: center;
-  justify-content: center;
-}
-
-@media screen and (max-width: 490px) {
-  .container {
-    height: 100vh;
-  }
-}
-
-.header {
-  display: block;
-}
-
-@media screen and (max-width: 490px) {
-  .header {
-    display: none;
-  }
-}
-
-.description {
-  justify-self: center;
-  font-family: 'Jet Brains Mono';
-  font-size: 16px;
-  line-height: 1.5;
-  font-weight: normal;
-}
-
-@media screen and (max-width: 490px) {
-  .description {
-    font-size: 20px;
-    line-height: 1.5;
-    font-weight: 700;
-    text-align: center;
-  }
-}
-
-.pic {
-  min-height: 240px;
-  min-width: 480px;
-}
-
-@media screen and (max-width: 490px) {
-  .pic {
-    min-height: 190px;
-    min-width: 350px;
-  }
-}
-
-.calContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  align-content: center;
-  justify-content: center;
-}
-
-@media screen and (max-width: 490px) {
-  .calContainer {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(2, 1fr);
-  }
-}
-
-.cal {
-  min-width: 120px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  align-content: center;
-  justify-content: center;
-}
